Kasir

+++ app/Http/Livewire/Sale/SaleIndex.php
@@ -6,7 +6,8 @@ use Livewire\Component;
 use Livewire\WithPagination;
 use Illuminate\Support\Facades\DB;
 use App\Models\Sale;
-
+use Illuminate\Support\Facades\Auth;
+use App\Helpers\Access;
 class SaleIndex extends Component
 {
     use WithPagination;
@@ -15,29 +16,50 @@ class SaleIndex extends Component
     public $perPage = 10;
     protected $paginationTheme = 'bootstrap';
 
-    protected $listeners = ['refreshSales' => '$refresh'];
+    protected $listeners = ['refreshSales' => '$refresh','deleteSale'];
+
 
     public function updatingSearch()
     {
         $this->resetPage();
     }
 
+
+    /**
+     * Dispatch a browser event to confirm deletion of a sale
+     *
+     * @param int $saleId The ID of the sale to be deleted
+     */
+    public function confirmDelete($saleId)
+    {
+        $this->dispatchBrowserEvent('confirm-delete', ['saleId' => $saleId]);
+    }
+
     public function deleteSale($saleId)
     {
+        if (!Access::can('destroy','sales')) 
+        {
+            $this->dispatchBrowserEvent('notify', [
+                'type' => 'error',
+                'message' => 'Anda tidak memiliki izin untuk menghapus penjualan'
+            ]);
+            return;
+        }
+
         try {
             DB::transaction(function () use ($saleId) {
-                $sale = Sale::findOrFail($saleId);
+                $sale = Sale::byCompany(auth()->user()->company_id)->findOrFail($saleId);
                 $sale->delete();
             });
 
             $this->dispatchBrowserEvent('notify', [
                 'type' => 'success',
-                'message' => 'Sale deleted successfully!'
+                'message' => 'Penjualan berhasil dihapus!'
             ]);
         } catch (\Exception $e) {
             $this->dispatchBrowserEvent('notify', [
                 'type' => 'error',
-                'message' => 'Error deleting sale: ' . $e->getMessage()
+                'message' => 'Error menghapus penjualan: ' . $e->getMessage()
             ]);
         }
     }
